refactor: rename deletion process to "asOwner" instead of "asUser"
