+++ Modules/Devices/src/Devices.Application/Identities/Commands/StartDeletionProcessAsOwner/Handler.cs
@@ -4,9 +4,9 @@ using Backbone.DevelopmentKit.Identity.Entities;
 using Backbone.Modules.Devices.Application.Infrastructure.Persistence.Repository;
 using MediatR;
 
-namespace Backbone.Modules.Devices.Application.Identities.Commands.StartDeletionProcessAsUser;
+namespace Backbone.Modules.Devices.Application.Identities.Commands.StartDeletionProcessAsOwner;
 
-public class Handler : IRequestHandler<StartDeletionProcessAsUserCommand, StartDeletionProcessAsUserResponse>
+public class Handler : IRequestHandler<StartDeletionProcessAsOwnerCommand, StartDeletionProcessAsOwnerResponse>
 {
     private readonly IIdentitiesRepository _identitiesRepository;
     private readonly IUserContext _userContext;
@@ -17,14 +17,14 @@ public class Handler : IRequestHandler<StartDeletionProcessAsUserCommand, StartD
         _userContext = userContext;
     }
 
-    public async Task<StartDeletionProcessAsUserResponse> Handle(StartDeletionProcessAsUserCommand request, CancellationToken cancellationToken)
+    public async Task<StartDeletionProcessAsOwnerResponse> Handle(StartDeletionProcessAsOwnerCommand request, CancellationToken cancellationToken)
     {
         var identity = await _identitiesRepository.FindByAddress(_userContext.GetAddress(), cancellationToken, true) ?? throw new NotFoundException(nameof(Identity));
 
-        var deletionProcess = identity.StartDeletionProcess(_userContext.GetDeviceId());
+        var deletionProcess = identity.StartDeletionProcessAsOwner(_userContext.GetDeviceId());
 
         await _identitiesRepository.Update(identity, cancellationToken);
 
-        return new StartDeletionProcessAsUserResponse(deletionProcess);
+        return new StartDeletionProcessAsOwnerResponse(deletionProcess);
     }
 }
