+++ Modules/Devices/src/Devices.Application/Identities/Commands/StartDeletionProcessAsOwner/StartDeletionProcessAsOwnerCommand.cs
@@ -1,0 +1,7 @@
+﻿using MediatR;
+
+namespace Backbone.Modules.Devices.Application.Identities.Commands.StartDeletionProcessAsOwner;
+
+public class StartDeletionProcessAsOwnerCommand : IRequest<StartDeletionProcessAsOwnerResponse>
+{
+}
