+++ Modules/Devices/src/Devices.Application/Identities/Commands/StartDeletionProcessAsOwner/StartDeletionProcessAsOwnerResponse.cs
@@ -1,11 +1,11 @@
 ﻿using Backbone.DevelopmentKit.Identity.ValueObjects;
 using Backbone.Modules.Devices.Domain.Entities.Identities;
 
-namespace Backbone.Modules.Devices.Application.Identities.Commands.StartDeletionProcessAsUser;
+namespace Backbone.Modules.Devices.Application.Identities.Commands.StartDeletionProcessAsOwner;
 
-public class StartDeletionProcessAsUserResponse
+public class StartDeletionProcessAsOwnerResponse
 {
-    public StartDeletionProcessAsUserResponse(IdentityDeletionProcess deletionProcess)
+    public StartDeletionProcessAsOwnerResponse(IdentityDeletionProcess deletionProcess)
     {
         Id = deletionProcess.Id;
         Status = deletionProcess.Status;
@@ -18,7 +18,9 @@ public class StartDeletionProcessAsUserResponse
     public string Id { get; set; }
     public DeletionProcessStatus Status { get; set; }
     public DateTime CreatedAt { get; set; }
+
     public DateTime ApprovedAt { get; set; }
     public DeviceId ApprovedByDevice { get; set; }
+
     public DateTime GracePeriodEndsAt { get; set; }
 }
