--- Modules/Devices/src/Devices.Application/Identities/Commands/StartDeletionProcessAsUser/StartDeletionProcessAsUserCommand.cs
@@ -1,7 +1,0 @@
-﻿using MediatR;
-
-namespace Backbone.Modules.Devices.Application.Identities.Commands.StartDeletionProcessAsUser;
-
-public class StartDeletionProcessAsUserCommand : IRequest<StartDeletionProcessAsUserResponse>
-{
-}
