+++ Modules/Devices/test/Devices.Application.Tests/Tests/Identities/Commands/StartDeletionProcessAsOwner/HandlerTests.cs
@@ -1,7 +1,7 @@
 ﻿using Backbone.BuildingBlocks.Application.Abstractions.Exceptions;
 using Backbone.BuildingBlocks.Application.Abstractions.Infrastructure.UserContext;
 using Backbone.DevelopmentKit.Identity.ValueObjects;
-using Backbone.Modules.Devices.Application.Identities.Commands.StartDeletionProcessAsUser;
+using Backbone.Modules.Devices.Application.Identities.Commands.StartDeletionProcessAsOwner;
 using Backbone.Modules.Devices.Application.Infrastructure.Persistence.Repository;
 using Backbone.Modules.Devices.Domain.Entities.Identities;
 using Backbone.UnitTestTools.Extensions;
@@ -10,7 +10,7 @@ using FluentAssertions;
 using Xunit;
 using static Backbone.UnitTestTools.Data.TestDataGenerator;
 
-namespace Backbone.Modules.Devices.Application.Tests.Tests.Identities.Commands.StartDeletionProcessAsUser;
+namespace Backbone.Modules.Devices.Application.Tests.Tests.Identities.Commands.StartDeletionProcessAsOwner;
 
 public class HandlerTests
 {
@@ -32,7 +32,7 @@ public class HandlerTests
         var handler = CreateHandler(mockIdentitiesRepository, fakeUserContext);
 
         // Act
-        var command = new StartDeletionProcessAsUserCommand();
+        var command = new StartDeletionProcessAsOwnerCommand();
         var response = await handler.Handle(command, CancellationToken.None);
 
         // Assert
@@ -68,11 +68,11 @@ public class HandlerTests
         var handler = CreateHandler(fakeIdentitiesRepository, fakeUserContext);
 
         // Act
-        var command = new StartDeletionProcessAsUserCommand();
+        var command = new StartDeletionProcessAsOwnerCommand();
         var acting = async () => await handler.Handle(command, CancellationToken.None);
 
         // Assert
-        acting.Should().AwaitThrowAsync<NotFoundException, StartDeletionProcessAsUserResponse>().Which.Message.Should().Contain("Identity");
+        acting.Should().AwaitThrowAsync<NotFoundException, StartDeletionProcessAsOwnerResponse>().Which.Message.Should().Contain("Identity");
     }
 
     private static Handler CreateHandler(IIdentitiesRepository identitiesRepository, IUserContext userContext)
